backend/asset-management-tool/requirement.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
CacheControl==0.12.6
cachetools==4.2.4
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Django==3.1.7
django-celery-beat==2.2.1
django-cors-headers==3.7.0
django-crispy-forms==1.13.0
django-filter==2.4.0
django-timezone-field==4.2.1
djangorestframework==3.11.2
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
firebase-admin==5.0.3
google-api-core==2.1.1
google-api-python-client==2.27.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.39.0
httplib2==0.20.1
humanize==3.11.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
msgpack==1.0.2
ndg-httpsclient==0.5.1
nepali-datetime==1.0.7
packaging==21.0
Pillow==8.3.2
prompt-toolkit==3.0.20
proto-plus==1.19.5
protobuf==3.19.0
psycopg2-binary==2.8.5
pusher-push-notifications==2.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
python-decouple==3.3
pytz==2021.1
redis==3.5.3
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
